banner

+++ bean/HomeBean.java
@@ -267,7 +267,7 @@ public class HomeBean {
             private Integer id;
             private String name;
             private String list_pic_url;
-            private Integer retail_price;
+            private Double retail_price;
 
             public Integer getId() {
                 return id;
@@ -293,11 +293,11 @@ public class HomeBean {
                 this.list_pic_url = list_pic_url;
             }
 
-            public Integer getRetail_price() {
+            public Double getRetail_price() {
                 return retail_price;
             }
 
-            public void setRetail_price(Integer retail_price) {
+            public void setRetail_price(Double retail_price) {
                 this.retail_price = retail_price;
             }
         }
@@ -314,7 +314,7 @@ public class HomeBean {
             private Integer id;
             private String name;
             private String list_pic_url;
-            private Integer retail_price;
+            private Double retail_price;
             private String goods_brief;
 
             public Integer getId() {
@@ -341,11 +341,11 @@ public class HomeBean {
                 this.list_pic_url = list_pic_url;
             }
 
-            public Integer getRetail_price() {
+            public Double getRetail_price() {
                 return retail_price;
             }
 
-            public void setRetail_price(Integer retail_price) {
+            public void setRetail_price(Double retail_price) {
                 this.retail_price = retail_price;
             }
 
@@ -383,7 +383,7 @@ public class HomeBean {
             private String pic_url;
             private Integer sort_order;
             private Integer is_show;
-            private Integer floor_price;
+            private Double floor_price;
             private String app_list_pic_url;
             private Integer is_new;
             private String new_pic_url;
@@ -445,11 +445,11 @@ public class HomeBean {
                 this.is_show = is_show;
             }
 
-            public Integer getFloor_price() {
+            public Double getFloor_price() {
                 return floor_price;
             }
 
-            public void setFloor_price(Integer floor_price) {
+            public void setFloor_price(Double floor_price) {
                 this.floor_price = floor_price;
             }
 
@@ -517,7 +517,7 @@ public class HomeBean {
             private String item_pic_url;
             private String subtitle;
             private Integer topic_category_id;
-            private Integer price_info;
+            private Double price_info;
             private String read_count;
             private String scene_pic_url;
             private Integer topic_template_id;
@@ -581,11 +581,11 @@ public class HomeBean {
                 this.topic_category_id = topic_category_id;
             }
 
-            public Integer getPrice_info() {
+            public Double getPrice_info() {
                 return price_info;
             }
 
-            public void setPrice_info(Integer price_info) {
+            public void setPrice_info(Double price_info) {
                 this.price_info = price_info;
             }
 
@@ -684,7 +684,7 @@ public class HomeBean {
                 private Integer id;
                 private String name;
                 private String list_pic_url;
-                private Integer retail_price;
+                private Double retail_price;
 
                 public Integer getId() {
                     return id;
@@ -710,11 +710,11 @@ public class HomeBean {
                     this.list_pic_url = list_pic_url;
                 }
 
-                public Integer getRetail_price() {
+                public Double getRetail_price() {
                     return retail_price;
                 }
 
-                public void setRetail_price(Integer retail_price) {
+                public void setRetail_price(Double retail_price) {
                     this.retail_price = retail_price;
                 }
             }
